feat: Add "Userga xabar yuborish" functionality to start scene
This commit adds the "Userga xabar yuborish" functionality to the start scene. When the user selects this option, they are prompted to enter a user ID and then a message. The message is then sent to the specified user via their Telegram ID. This feature enhances communication capabilities within the application.